feat: add no data SVG asset and implement empty/error states in HomeScreen
- Added a new SVG asset for the no data state.
- Updated HomeScreen to display an empty state when there are no tickets.
- Implemented an error state with retry functionality when ticket loading fails.
- Refactored ticket card to use double for betAmount instead of int for better precision.
- Enhanced user experience with SnackBar for error notifications and improved date formatting.

+++ lib/modules/home/screens/widgets/ticket_card.dart
@@ -7,7 +7,7 @@ enum Status { pending, won, lost }
 class TicketCard extends StatelessWidget {
   final String ticketId;
   final String bookingCode;
-  final int betAmount;
+  final double betAmount;
   final DateTime trackedAt;
   final String description;
   final double totalOdds;
